+++ source/ProcessManager.Orchestrations/Processes/BRS_021/ElectricalHeatingCalculation/Databricks/SqlStatements/CalculatedMeasurementsColumnNames.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_021.ElectricalHeatingCalculation.SqlStatements;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_021.ElectricalHeatingCalculation.Databricks.SqlStatements;
 
 internal static class CalculatedMeasurementsColumnNames
 {
